requirements.txt
Flask==2.1.2
azure_storage==0.37.0
PyYAML==6.0
